dsp-client/src/main/java/com/flipkart/dsp/client/request/GetRequestByRequestId.java
package com.flipkart.dsp.client.request;

import com.fasterxml.jackson.databind.JavaType;
import com.flipkart.dsp.client.misc.AbstractDSPServiceRequest;
import com.flipkart.dsp.client.DSPServiceClient;
import com.flipkart.dsp.entities.request.Request;
import com.flipkart.dsp.utils.JsonUtils;
import com.ning.http.client.RequestBuilder;

/**
 * +
 */
public class GetRequestByRequestId extends AbstractDSPServiceRequest<Request> {
    private Long requestId;

    public GetRequestByRequestId(DSPServiceClient dspServiceClient, Long requestId) {
        super(dspServiceClient);
        this.requestId = requestId;
    }

    @Override
    protected String getMethod() {
        return "GET";
    }

    @Override
    protected String getPath() {
        return "/v2/requests/" + requestId;
    }

    @Override
    protected JavaType getReturnType() {
        return JsonUtils.DEFAULT.mapper.getTypeFactory().constructType(Request.class);
    }

    @Override
    protected RequestBuilder buildRequest(RequestBuilder requestBuilder) {
        return requestBuilder;
    }
}
